requirements.txt
fastapi
uvicorn[standard]
gunicorn
ultralytics
pillow
python-multipart
torch==2.2.2
torchvision==0.17.2
numpy<2